+++ files/basic.rt
@@ -1,0 +1,6 @@
+A   0.2	                                    255,255,255
+
+C   -50,0,20        0,0,0       70
+L   5,5,-10                    1.0         255,255,255
+
+sp  0,0,20                      20          255,0,0
