+++ includes/minirt.h
@@ -9,120 +9,8 @@
 # include <math.h>
 # include "libft.h"
 # include "key.h"
-# include "minirt_error.h"
-
-# ifdef __linux__
-#  define IS_LINUX 1
-#  define OS "LINUX"
-# else
-#  define IS_LINUX 0
-#  define OS "MAC"
-# endif
-
-# define ERROR_MLX	1
-# define ERROR_WIN	2
-# define FRAME_RATE	60
-
-# define PLANE	1
-# define PLAN	1
-# define SPHERE	2
-# define CYLIND	3
-# define C		3
-
-# define WIDTH	1280
-# define HEIGHT	720
-
-# define READ_SIZE 1024
-
-typedef struct s_size
-{
-	double	w; // width
-	double	h; // height
-	double	d; // depth
-}	t_size;
-
-typedef struct s_color
-{
-	int	r;  // red
-	int	g;  // green
-	int	b;  // blue
-}	t_color;
-
-typedef struct s_vtr
-{
-	double	x;
-	double	y;
-	double	z;
-	double	mag;
-}	t_vtr;
-
-typedef struct s_ray
-{
-	t_vtr	a;
-	t_vtr	b;
-	t_vtr	l;
-}	t_ray;
-
-typedef struct s_img {
-	void	*ptr;
-	char	*addr;
-	int		bpp;
-	int		lh;
-	int		endian;
-}	t_img;
-
-typedef struct s_ambient
-{
-	double	ratio;	//* range[0.0-1.0];
-	t_color	color;	//* range[0-255];
-}	t_amb;
-
-typedef struct s_camara
-{
-	t_vtr	pos;
-	t_vtr	norm;	//* range[-1-1];
-	int		fov;	//* horizountal field view in degree range[0-180];
-	t_vtr	lookat;
-	t_vtr	up;
-	double	length;
-	double	horz;
-	double	ratio;
-	t_vtr	proj_c;	// norm_vector in projection center screen
-	t_vtr	proj_u;	// norm_vector in projection screen u
-	t_vtr	proj_v;	// norm_vector in projection screen v
-} 	t_cam;
-
-typedef struct s_light
-{
-	t_vtr	pos;
-	double	bright; //* range[0.0-1.0];
-	t_color	color;
-}	t_lht;
-
-typedef struct s_obj {
-	int		idx;
-	int		type;
-	char	*name;
-	t_vtr	pos;
-	t_vtr	norm;
-	t_color	color;
-	t_size	size;
-	t_img	img;
-}	t_obj;
-
-typedef struct s_data
-{
-	void	*mlx;
-	void	*win;
-	int		w;
-	int		h;
-	int		frame;
-	t_obj	scene;
-	t_amb	amb;
-	t_cam	cam;
-	t_list	*lht;
-	t_list	*objs;
-}	t_data;
+# include "error.h"
+# include "data.h"
 
 int	rt_close(t_data *data, int code);
 int	error_exit(t_data *data, int code);
@@ -132,25 +20,38 @@ int rt_setup(t_data *data, char *filename);
 // * parsing_intput
 t_data	parsing_input(int argc, char **argv);
 void	ambient_initialise(t_data *data, char **object);
-void	camara_initialise(t_data *data, char **object);
-void	light_initialise(t_data* data, char **object);
-void	sephere_initialise(t_data *data, char **object);
+t_obj	*object_initialise(int type);
 void	plane_initialise(t_data *data, char **object);
 void	cylinder_inititialize(t_data *data, char **object);
 
-// render
+// * render
 void	render_objects(t_data *data);
+int 	render_scene(t_data *data);
 
-// obj
+// * camera
+void	cam_initialise(t_data *data, char **object);
+void	cam_geometry(t_cam *cam);
+t_ray	cam_ray(t_cam cam, float proj_x, float proj_y);
+
+// * light
+void	lht_initialise(t_data* data, char **object);
+void	lht_illuminated(t_lht lht, t_ints *ints);
+
+// * obj
 void	set_objs(t_data *data);
 t_obj	*new_obj(t_data *data);
 void	free_obj(t_obj *obj);
-
-int	sphere_inter(t_ray ray, t_vtr *vinter);
-
+void	scene_initialise(t_data *data);
+void	sphere_initialise(t_data *data, char **object);
+void	sphere_inter(t_ray ray, t_ints *ints);
 t_obj	*set_plain_img(t_data *data, t_obj *obj);
 
-// utils
+
+// * unit
+t_ray	rayset(t_vtr a, t_vtr b);
+
+// * utils
+int		color_to_int(t_color color);
 int		rgb_to_int(int r, int g, int b);
 void	pixel_put_img(t_img *img, int x, int y, int color);
 t_color	set_color(int r, int g, int b);
@@ -164,11 +65,6 @@ t_vtr	vtrsub(t_vtr a, t_vtr b);
 t_vtr	vtrscale(t_vtr v, double s);
 double	vtrdot(t_vtr a, t_vtr b);
 t_vtr	vtrcross(t_vtr a, t_vtr b);
-
-void	camgeometry(t_cam *cam);
-
-t_ray	rayset(t_vtr a, t_vtr b);
-t_ray	raygen(t_cam cam, float proj_x, float proj_y);
 
 // * utils from sharnvon..
 void	exit_error(char *str);
