+++ srcs/parsing/parsing_input.c
@@ -14,35 +14,18 @@
 #include "math.h"
 #include <string.h>
 
-/* 
+/*
 TODO utility functions that must add into header files
 */
-
-void	exit_error(char *str);
-void	free_twopointer_char(char **ptr);
-char	*ft_strjoin_pro(char *str1, char *str2);
-int		character_count(char *str, char c);
-double	ato_double(char *str);
-t_color	ato_tcolor(char *str);
-t_vtr	ato_tvector(char *str);
-int		tvector_inrange(t_vtr vactor, double min, double max);
-t_size	size_initialise(double height, double width, double depth);
 
 static int	file_checking(char *file);
 static void	get_input(t_data *data, int fd);
 static int	identifier_checking(char **object, int index);
 static void	object_lexering(t_data *data, char **object, int identifier);
 
-void	ambient_initialise(t_data *data, char **object);
-void	camara_initialise(t_data *data, char **object);
-void	light_initialise(t_data* data, char **object);
-void	sephere_initialise(t_data *data, char **object);
-void	plane_initialise(t_data *data, char **object);
-void	cylinder_inititialize(t_data *data, char **object);
-
 /*
 * [function parsing input from the file and conveting into struct data]
-* => [return] : 
+* => [return] :
 * => [exit] :
 TODO check lists amount of "data.lht" | 0 < lht_amount < 3 |
 TODO check lists amount of "data.objs" | 0 < objs_amount |
@@ -103,7 +86,7 @@ static void	get_input(t_data *data, int fd)
 		}
 		file = ft_strjoin_pro(file, buff);
 	}
-	
+
 	char	**object;
 	int		index;
 	int		identifier;
@@ -124,7 +107,7 @@ static void	get_input(t_data *data, int fd)
 		identifier = identifier_checking(object, -1);
 		if (identifier < 0)
 			exit_error ("minirt: invalid identifier...");
-		
+
 		// * lexering
 		object_lexering(data, object, identifier);
 		free_twopointer_char(object);
@@ -142,11 +125,11 @@ void	object_lexering(t_data *data, char **object, int identifier)
 	if (!ft_strncmp(object[identifier], "A", 2))
 		ambient_initialise(data, object);
 	else if (!ft_strncmp(object[identifier], "C", 2))
-		camara_initialise(data, object);// printf("%s\n", "go to fucntion C");
+		cam_initialise(data, object);// printf("%s\n", "go to fucntion C");
 	else if (!ft_strncmp(object[identifier], "L", 2))
-		light_initialise(data, object);//printf("%s\n", "go to fucntion L");
+		lht_initialise(data, object);//printf("%s\n", "go to fucntion L");
 	else if (!ft_strncmp(object[identifier], "sp", 3))
-		sephere_initialise(data, object);//printf("%s\n", "go to fucntion sp");
+		sphere_initialise(data, object);//printf("%s\n", "go to fucntion sp");
 	else if (!ft_strncmp(object[identifier], "pl", 3))
 		plane_initialise(data, object);//printf("%s\n", "go to fucntion ol");
 	else if (!ft_strncmp(object[identifier], "cy", 3))
