--- srcs/utils/debug.c
@@ -1,7 +1,0 @@
-#include "minirt.h"
-
-void	print_obj(t_obj *obj)
-{
-	// printf("code = %d\n", obj->type);
-	
-}
