+++ srcs/utils/pixel.c
@@ -1,8 +1,12 @@
 #include "minirt.h"
 
+int	color_to_int(t_color color)
+{
+	return (color.r << 16 | color.g << 8 | color.b);
+}
+
 int	rgb_to_int(int r, int g, int b)
 {
-	// return ((r & 0xff) << 16 | (g & 0xff) << 8 | (b & 0xff));
 	return (r << 16 | g << 8 | b);
 }
